// front/src/model/GoodsCard.tsx
export interface GoodsCard{
    id_g: number;
    title: string;
    price: number;
    firstImage: string;
    article: string;
    quantity: number;
    description: string;
    mark: number;
    composition: string;
    state: string;
    typeName: string;
    shopName: string;
    image: string[];
    size: string[];
}